move the updates disabled window logic to submodule
